Refactor DOM organization
