+++ src/index.html
@@ -198,7 +198,7 @@
     </section>
     <footer>
         <input name="search" type="submit" class="hidden search-submit-btn" id="searchFavesSubmitBtn">
-        <input type="text" role="searchbox" placeholder= "Search favorites" name="searchFavorites" class="hidden search-field" id="searchFavesByName">
+        <input type="text" role="searchbox" placeholder= "Search favorites" name="searchFavorites" class="hidden search-favorites-field" id="searchFavesByName">
     </footer>
     <article class="popup-container">
       <span class="hidden" id="popup" aria-label="popup message">
